Always update Packet's addresses
This is useful when one pings multicast addresses to see the actual
IPs that replied. There exist multicast addresses for both IPv4 and
IPV6 but the lib does not provide the replies address.

The output of ping without this patch is:

$ sudo ping --privileged ff02::1%ens3
PING ff02::1%ens3 (ff02::1%ens3):
24 bytes from ff02::1%ens3: icmp_seq=0 time=1.504454ms ttl=64
24 bytes from ff02::1%ens3: icmp_seq=0 time=1.553965ms ttl=64 (DUP!)
24 bytes from ff02::1%ens3: icmp_seq=0 time=1.580495ms ttl=64 (DUP!)
24 bytes from ff02::1%ens3: icmp_seq=0 time=1.596156ms ttl=64 (DUP!)

And with this patch:

$ sudo ./ping --privileged ff02::1%ens3
PING ff02::1%ens3 (ff02::1%ens3):
32 bytes from fe80::546f:f0ff:fe0e:a%ens3: icmp_seq=0 time=1.256761ms ttl=64
32 bytes from fe80::546f:f0ff:fe0e:29%ens3: icmp_seq=0 time=1.396183ms ttl=64 (DUP!)
32 bytes from fe80::546f:f0ff:fe0e:8%ens3: icmp_seq=0 time=1.447454ms ttl=64 (DUP!)
32 bytes from fe80::546f:f0ff:fe0e:e%ens3: icmp_seq=0 time=1.460754ms ttl=64 (DUP!)

Which provide more useful informations.

Signed-off-by: Gregory Detal <gregory.detal@tessares.net>